-> Split various convenience methods out of metadata_helpers into a new module, metadata_utils -> Add 'solr_helpers' module which provides convenience methods for interfacing with Apache Solr -> Added convenience method get_cohort_cases to Cohort model -> Overhaul of DICOM section of the file_helpers.cohort_files method
